requirements.txt
streamlit>=1.30
pandas
numpy
scikit-learn
xgboost
joblib
requests
python-dotenv
altair
